Fix #103 : Treenode width not well calculated when component hidden then displayed
