+++ src/ferui-components/tree-view/tree-node-component.ts
@@ -9,11 +9,22 @@ import {
   ElementRef,
   ViewChild,
   ChangeDetectionStrategy,
-  ChangeDetectorRef
+  ChangeDetectorRef,
+  Self,
+  OnDestroy,
+  DoCheck
 } from '@angular/core';
-import { TreeViewEventType, TreeNodeDataRetriever, PagedTreeNodeDataRetriever, TreeViewColorTheme } from './interfaces';
+import {
+  TreeViewEventType,
+  TreeNodeDataRetriever,
+  PagedTreeNodeDataRetriever,
+  TreeViewColorTheme,
+  TreeViewConfiguration
+} from './interfaces';
 import { TreeNode, TreeNodeEvent } from './internal-interfaces';
 import { FuiTreeViewUtilsService } from './tree-view-utils-service';
+import { DomObserver, ObserverInstance } from '../utils/dom-observer/dom-observer';
+import { ScrollbarHelper } from '../utils/scrollbar-helper/scrollbar-helper.service';
 
 @Component({
   selector: 'fui-tree-node',
@@ -49,19 +60,16 @@ import { FuiTreeViewUtilsService } from './tree-view-utils-service';
   `,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class FuiTreeNodeComponent<T> implements OnInit {
+export class FuiTreeNodeComponent<T> implements OnInit, OnDestroy, DoCheck {
   @Output() onNodeEvent: EventEmitter<TreeNodeEvent<T>> = new EventEmitter<TreeNodeEvent<T>>();
 
   @Input() node: TreeNode<T>;
-
   @Input() theme: TreeViewColorTheme;
-
   @Input() dataRetriever: TreeNodeDataRetriever<T> | PagedTreeNodeDataRetriever<T>;
+  @Input() treeviewConfig: TreeViewConfiguration;
 
   @HostBinding('class') themeClass;
-
   @HostBinding('class.fui-tree-node-component') nodeComponent: boolean = true;
-
   @HostBinding('class.borders') @Input() borders: boolean = false;
 
   @ViewChild('nodetree', { read: ElementRef }) nodeTreeElement: ElementRef;
@@ -69,11 +77,18 @@ export class FuiTreeNodeComponent<T> implements OnInit {
   // Hierarchical level to show parent-child relationship
   level: number = 0;
   // Indicates node can be expanded
-  hasChildren: boolean;
+  hasChildren: boolean = false;
   // left padding dependent of node tree hierarchical level
   padding: number;
 
-  constructor(private cd: ChangeDetectorRef, private treeViewUtils: FuiTreeViewUtilsService) {}
+  private domObservers: ObserverInstance[] = [];
+
+  constructor(
+    @Self() private element: ElementRef,
+    private scrollbarHelper: ScrollbarHelper,
+    private cd: ChangeDetectorRef,
+    private treeViewUtils: FuiTreeViewUtilsService
+  ) {}
 
   /**
    * Initiates Tree Node component by setting its hierarchical level
@@ -88,13 +103,20 @@ export class FuiTreeNodeComponent<T> implements OnInit {
     }
     this.dataRetriever.hasChildNodes(this.node.data).then((hasChildren: boolean) => {
       this.hasChildren = hasChildren;
-      this.padding = this.calculatePadding();
-      this.cd.markForCheck();
-      // At this particular time the view isn't updated fast enough so we calculate the node width to include:
-      // Node text width, padding and if icons exists take a value of 16 into consideration
-      const iconPadding = this.hasChildren ? 16 : 0;
-      this.treeViewUtils.virtualScrollerWidth = this.getNodeWidth() + iconPadding + this.padding;
+      this.setVirtualScrollerWidth();
     });
+
+    // Once the node is visible on screen we need to re-calculate the virtual scroller width.
+    const headerViewport: Element = this.element.nativeElement;
+    this.domObservers.push(
+      DomObserver.observe(headerViewport, entities => {
+        entities.forEach(entity => {
+          if (entity.isIntersecting) {
+            this.setVirtualScrollerWidth();
+          }
+        });
+      })
+    );
   }
 
   /**
@@ -103,6 +125,11 @@ export class FuiTreeNodeComponent<T> implements OnInit {
    */
   ngDoCheck(): void {
     this.cd.markForCheck();
+  }
+
+  ngOnDestroy() {
+    this.domObservers.forEach(observerInstance => DomObserver.unObserve(observerInstance));
+    this.domObservers = undefined;
   }
 
   /**
@@ -152,6 +179,24 @@ export class FuiTreeNodeComponent<T> implements OnInit {
    */
   calculatePadding(): number {
     return this.hasChildren ? this.level * 20 + 10 : this.level * 20 + 30;
+  }
+
+  /**
+   * Set virtual scroller width depending on node width or config width.
+   */
+  private setVirtualScrollerWidth() {
+    this.padding = this.calculatePadding();
+    // At this particular time the view isn't updated fast enough so we calculate the node width to include:
+    // Node text width, padding and if icons exists take a value of 16 into consideration
+    const iconPadding = this.hasChildren ? 16 : 0;
+    const currentNodeWidth: number = this.getNodeWidth() + iconPadding + this.padding;
+    const configWidth: number = this.treeviewConfig ? parseInt(this.treeviewConfig.width, 10) : 0;
+    // For users with scrollbars visible, we need to take the scrollbar width into account.
+    const scrollbarWidth: number = this.scrollbarHelper.getWidth();
+    this.treeViewUtils.virtualScrollerWidth =
+      configWidth > currentNodeWidth ? (this.borders ? configWidth : configWidth - (40 + scrollbarWidth)) : currentNodeWidth;
+    // 40 = 2 times padding of 20px.
+    this.cd.markForCheck();
   }
 
   /**
