+++ src/ferui-components/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './conditional/index';
 export * from './i18n/index';
 export * from './dom-observer/dom-observer';
+export * from './scrollbar-helper/scrollbar-helper.service';
